+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Headers:
   betters and lets calling application disambiguate on the type of error. 
 - Errors no longer don't duplicate the inner error's message and they are all
   lowercase, as prescribed by the Rust API guidelines.
+- Collapse `ClientBuildError::BuildError` into `ClientBuildError`. The enum only
+  had one member.
 
 ### Fixed